ジャンル検索、範囲検索メニューの部分を切り離し一つのコンポーネントにした
